Header files now appear under "libserial" subfolder.
Updated code and examples so that LibSerial header files are now
expected to be installed under "libserial" subfolder of the include
directory specified during installation. The goal is the keep all
LibSerial headers under a separate folder so they do not conflict with
similarly named headers from other packages.
