+++ examples/example_project/CMakeLists.txt
@@ -21,7 +21,6 @@ INCLUDE(GNUInstallDirs)
 #
 SET(THREADS_HAVE_PTHREAD_ARG 1)
 FIND_PACKAGE(PythonLibs REQUIRED)
-#FIND_PACKAGE(SIP REQUIRED)
 FIND_PACKAGE(Threads REQUIRED)
 
 #
