+++ examples/example_project/example_project.cpp
@@ -2,8 +2,8 @@
  *  @example example_project.cpp
  */
 
-#include <SerialPort.h>
-#include <SerialStream.h>
+#include <libserial/SerialPort.h>
+#include <libserial/SerialStream.h>
 
 #include <iostream>
