+++ src/SerialStream.cpp
@@ -30,7 +30,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE            *
  * POSSIBILITY OF SUCH DAMAGE.                                                *
  *****************************************************************************/
-#include "SerialStream.h"
+#include "libserial/SerialStream.h"
 
 #include <cassert>
