+++ src/libserial/SerialStream.h
@@ -1,5 +1,5 @@
 /******************************************************************************
- * @file SerialStreamBuf.h                                                    *
+ * @file SerialStream.h                                                       *
  * @copyright (C) 2004-2018 LibSerial Development Team. All rights reserved.  *
  * crayzeewulf@gmail.com                                                      *
  *                                                                            *
@@ -33,41 +33,62 @@
 
 #pragma once
 
-#include <SerialPortConstants.h>
+#include <libserial/SerialPortConstants.h>
+#include <libserial/SerialStreamBuf.h>
 
+#include <istream>
 #include <memory>
-#include <streambuf>
-#include <vector>
 
 namespace LibSerial
 {
     /**
-     * @brief SerialStreamBuf is the streambuf subclass used by SerialStream.
-     *        This subclass takes care of opening the serial port file in the
-     *        required modes and providing the corresponding file descriptor
-     *        to SerialStream so that various parameters associated with the
-     *        serial port can be set. Several features of this streambuf class
-     *        resemble those of std::filebuf, however this class it not made a
-     *        subclass of filebuf because we need access to the file descriptor
-     *        associated with the serial port and the standard filebuf does not
-     *        provide access to it.
+     * @brief SerialStream is a stream class for accessing serial ports on
+     *        POSIX operating systems. A lot of the functionality of this class
+     *        has been obtained by looking at the code of libserial package by
+     *        Linas Vepstas, (linas@linas.org) and the excellent document on
+     *        serial programming by Michael R. Sweet. This document can be
+     *        found at
+     *        <ahref="http://www.easysw.com/~mike/serial/serial.html">
+     *        http://www.easysw.com/~mike/serial/serial.html</a>.
+     *        The libserial package can be found at
+     *        <ahref="http://www.linas.org/serial/">
+     *        http://www.linas.org/serial/</a>.
+     *        This class allows one to set various parameters of a serial
+     *        port and then access it like a simple fstream. (In fact, that
+     *        is exactly what it does!) It sets the parameters of the
+     *        serial port by maintaining a file descriptor for the port and
+     *        uses the basic_fstream functions for the IO. We have not
+     *        implemented any file locking yet but it will be added soon.
      *
-     *        At present, this class uses unbuffered I/O and all calls
-     *        to setbuf() will be ignored.
+     *        Make sure you read the documentation of the standard fstream
+     *        template before using this class because most of the
+     *        functionality is inherited from fstream. Also, a lot of
+     *        information about the various system calls used in the
+     *        implementation can also be found in the Single Unix
+     *        Specification (Version 2). A copy of this document can be
+     *        obtained from <a href="http://www.UNIX-systems.org/">
+     *        http://www.UNIX-systems.org/</a>. We will refer to this
+     *        document as SUS-2.
      */
-    class SerialStreamBuf : public std::streambuf
+    class SerialStream : public std::iostream
     {
     public:
 
         /**
-         * @brief Default Constructor.
+         * @brief Default Contructor.
+         *        Creates a new SerialStream object but does not open it.
+         *        The Open() method will need to be called explicitly on
+         *        the object to communicate with the serial port.
          */
-        explicit SerialStreamBuf() ;
+        explicit SerialStream() ;
 
         /**
-         * @brief Constructor that allows a SerialStreamBuf instance to be 
+         * @brief Constructor that allows a SerialStream instance to be
          *        created and opened, initializing the corresponding
          *        serial port with the specified parameters.
+         *        Suggested by Witek Adamus (wit3k):
+         *        https://sourceforge.net/tracker/index.php?func=detail&aid=2137885&group_id=9432&atid=359432
+         *
          * @param fileName The file name of the serial stream.
          * @param baudRate The communications baud rate.
          * @param characterSize The size of the character buffer for
@@ -76,39 +97,43 @@ namespace LibSerial
          * @param stopBits The number of stop bits for the serial stream.
          * @param flowControlType The flow control type for the serial stream.
          */
-        explicit SerialStreamBuf(const std::string&   fileName,
-                                 const BaudRate&      baudRate        = BaudRate::BAUD_DEFAULT,
-                                 const CharacterSize& characterSize   = CharacterSize::CHAR_SIZE_DEFAULT,
-                                 const FlowControl&   flowControlType = FlowControl::FLOW_CONTROL_DEFAULT,
-                                 const Parity&        parityType      = Parity::PARITY_DEFAULT,
-                                 const StopBits&      stopBits        = StopBits::STOP_BITS_DEFAULT) ;
+        explicit SerialStream(const std::string&   fileName,
+                              const BaudRate&      baudRate        = BaudRate::BAUD_DEFAULT,
+                              const CharacterSize& characterSize   = CharacterSize::CHAR_SIZE_DEFAULT,
+                              const FlowControl&   flowControlType = FlowControl::FLOW_CONTROL_DEFAULT,
+                              const Parity&        parityType      = Parity::PARITY_DEFAULT,
+                              const StopBits&      stopBits        = StopBits::STOP_BITS_DEFAULT) ;
 
         /**
-         * @brief Default Destructor for a SerialStreamBuf object. Closes the
-         *        serial port associated with mFileDescriptor if open.
+         * @brief Default Destructor for a SerialStream object
+         *        Closes the stream associated with mFileDescriptor, and
+         *        also closes the serial port if open.
+         *        Remaining actions are accomplished by the fstream destructor.
          */
-        virtual ~SerialStreamBuf() ;
+        virtual ~SerialStream() ;
 
         /**
-         * @brief Copy construction is disallowed.
+         * @brief Prevents copying of objects of this class by declaring the copy
+         *        constructor private. This method is never defined.
          */
-        SerialStreamBuf(const SerialStreamBuf& otherSerialStreamBuf) = delete ;
+        SerialStream(const SerialStream& otherSerialStream) = delete;
 
         /**
          * @brief Move construction is disallowed.
          */
-        SerialStreamBuf(const SerialStreamBuf&& otherSerialStreamBuf) = delete ;
+        SerialStream(const SerialStream&& otherSerialStream) = delete;
 
         /**
-         * @brief Copy assignment is disallowed.
+         * @brief Prevents copying of objects of this class by declaring the
+         *        assignment operator private. This method is never defined.
          */
-        SerialStreamBuf& operator=(const SerialStreamBuf& otherSerialStreamBuf) = delete ;
+        SerialStream& operator=(const SerialStream& otherSerialStream) = delete;
 
         /**
-         * @brief Move assignment is disallowed.
+         * @brief Move assignment is not allowed.
          */
-        SerialStreamBuf& operator=(const SerialStreamBuf&& otherSerialStreamBuf) = delete ;
-
+        SerialStream& operator=(const SerialStream&& otherSerialStream) = delete;
+    
         /**
          * @brief Opens the serial port associated with the specified
          *        file name and the specified mode.
@@ -155,12 +180,7 @@ namespace LibSerial
          * @brief Determines if the serial port is open for I/O.
          * @return Returns true iff the serial port is open.
          */
-        bool IsOpen() const ;
-
-        /**
-         * @brief Sets all serial port paramters to their default values.
-         */
-        void SetDefaultSerialPortParameters() ;
+        bool IsOpen() ;
 
         /**
          * @brief Sets the baud rate for the serial port to the specified value
@@ -172,7 +192,7 @@ namespace LibSerial
          * @brief Gets the current baud rate for the serial port.
          * @return Returns the baud rate.
          */
-        BaudRate GetBaudRate() const ;
+        BaudRate GetBaudRate() ;
 
         /**
          * @brief Sets the character size for the serial port.
@@ -182,9 +202,9 @@ namespace LibSerial
 
         /**
          * @brief Gets the character size being used for serial communication.
-         * @return Returns the current character size. 
+         * @return Returns the current character size.
          */
-        CharacterSize GetCharacterSize() const ;
+        CharacterSize GetCharacterSize() ;
 
         /**
          * @brief Sets flow control for the serial port.
@@ -196,7 +216,7 @@ namespace LibSerial
          * @brief Gets the current flow control setting.
          * @return Returns the flow control type of the serial port.
          */
-        FlowControl GetFlowControl() const ;
+        FlowControl GetFlowControl() ;
 
         /**
          * @brief Sets the parity type for the serial port.
@@ -208,7 +228,7 @@ namespace LibSerial
          * @brief Gets the parity type for the serial port.
          * @return Returns the parity type.
          */
-        Parity GetParity() const ;
+        Parity GetParity() ;
 
         /**
          * @brief Sets the number of stop bits to be used with the serial port.
@@ -220,7 +240,7 @@ namespace LibSerial
          * @brief Gets the number of stop bits currently being used by the serial
          * @return Returns the number of stop bits.
          */
-        StopBits GetStopBits() const ;
+        StopBits GetStopBits() ;
 
         /**
          * @brief Sets the minimum number of characters for non-canonical reads.
@@ -235,20 +255,20 @@ namespace LibSerial
          * @return Returns the minimum number of characters for
          *         non-canonical reads.
          */
-        short GetVMin() const ;
+        short GetVMin() ;
 
-        /** 
+        /**
          * @brief Sets character buffer timeout for non-canonical reads in deciseconds.
          * @param vtime The timeout value in deciseconds to be set.
          * @return Returns the character buffer timeout for non-canonical reads in deciseconds.
          */
         void SetVTime(short vtime) ;
 
-        /** 
+        /**
          * @brief Gets the current timeout value for non-canonical reads in deciseconds.
-         * @return Returns the character buffer timeout for non-canonical reads in deciseconds. 
+         * @return Returns the character buffer timeout for non-canonical reads in deciseconds.
          */
-        short GetVTime() const ;
+        short GetVTime() ;
 
         /**
          * @brief Sets the DTR line to the specified value.
@@ -261,7 +281,7 @@ namespace LibSerial
          * @brief Gets the status of the DTR line.
          * @return Returns true iff the status of the DTR line is high.
          */
-        bool GetDTR() const ;
+        bool GetDTR() ;
 
         /**
          * @brief Set the RTS line to the specified value.
@@ -274,7 +294,7 @@ namespace LibSerial
          * @brief Get the status of the RTS line.
          * @return Returns true iff the status of the RTS line is high.
          */
-        bool GetRTS() const ;
+        bool GetRTS() ;
 
         /**
          * @brief Get the status of the CTS line.
@@ -292,7 +312,7 @@ namespace LibSerial
          * @brief Gets the serial port file descriptor.
          * @return Returns the serial port file descriptor.
          */
-        int GetFileDescriptor() const ;
+        int GetFileDescriptor() ;
 
         /**
          * @brief Gets the number of bytes available in the read buffer.
@@ -303,121 +323,22 @@ namespace LibSerial
         /**
          * @brief Gets a list of available serial ports.
          * @return Returns a std::vector of std::strings with the name of
-         *         each available serial port. 
+         *         each available serial port.
+         *
+         * @todo Consider making this a static member function.
          */
-        std::vector<std::string> GetAvailableSerialPorts() const ;
+        std::vector<std::string> GetAvailableSerialPorts() ;
+
 
     protected:
-
-        /**
-         * @brief Performs an operation that is defined separately for each
-         *        class derived from streambuf. The default behavior is to
-         *        do nothing if gptr() is non-null and gptr()!=egptr().
-         *        Also, setbuf(0, 0) usually means unbuffered I/O and
-         *        setbuf(p, n) means use p[0]...p[n-1] to hold the buffered
-         *        characters. In general, this method implements the
-         *        subclass's notion of getting memory for the buffered
-         *        characters. 
-         *
-         *        In the case of SerialStreamBuffer, we want to keep using
-         *        unbuffered I/O. Hence, using this method has no effect at
-         *        present.
-         *
-         *
-         * @param character Pointer to the character buffer to write to the serial port.
-         * @param numberOfBytes The number of characters to write to the serial port.
-         * @return Returns a pointer to this streambuf object.
-         *
-         */
-        virtual std::streambuf* setbuf(char_type* character, 
-                                       std::streamsize numberOfBytes) override ;
-
-        /**
-         * @brief Writes up to n characters from the character sequence at 
-         *        char s to the serial port associated with the buffer.
-         * @param character Pointer to the character buffer to write to the serial port.
-         * @param numberOfBytes The number of characters to write to the serial port.
-         * @return Returns the number of characters that were successfully
-         *         written to the serial port. 
-         */
-        virtual std::streamsize xsputn(const char_type* character, 
-                                       std::streamsize numberOfBytes) override ;
-           
-        /**
-         * @brief Reads up to n characters from the serial port and returns
-         *        them through the character array located at s.
-         * @param character Pointer to the character buffer to write to the serial port.
-         * @param numberOfBytes The number of characters to write to the serial port.
-         * @return Returns the number of characters actually read from the
-         *         serial port. 
-         */
-        virtual std::streamsize xsgetn(char_type* character, 
-                                       std::streamsize numberOfBytes) override ;
-
-        /**
-         * @brief Writes the specified character to the associated serial port.
-         * @param character The character to be written to the serial port.
-         * @return Returns the character. 
-         */
-        virtual int_type overflow(const int_type character) override ;
-
-        /**
-         * @brief Reads and returns the next character from the associated
-         *        serial port if one otherwise returns traits::eof(). This
-         *        method is used for buffered I/O while uflow() is called
-         *        for unbuffered I/O.
-         * @return Returns the next character from the serial port.
-         */
-        virtual int_type underflow() override ;
-
-        /**
-         * @brief Reads and returns the next character from the associated
-         *        serial port if one otherwise returns traits::eof(). This
-         *        method is used for unbuffered I/O while underflow() is
-         *        called for unbuffered I/O.
-         * @return Returns the next character from the serial port.  
-         */
-        virtual int_type uflow() override ;
-
-        /**
-         * @brief This function is called when a putback of a character
-         *        fails. This must be implemented for unbuffered I/O as all
-         *        streambuf subclasses are required to provide putback of
-         *        at least one character.
-         * @param character The character to putback.
-         * @return Returns The character iff successful, otherwise eof to signal an error.
-         */
-        virtual int_type pbackfail(const int_type character) override ;
-
-        /**
-         * @brief Checks whether input is available on the port.
-         *        If you call \c SerialStream::in_avail, this method will
-         *        be called to check for available input.
-         *        \code
-         *        while(serial_port.rdbuf()->in_avail() > 0)
-         *        {
-         *            serial_port.get(ch) ;
-         *            ...
-         *        }
-         *        \endcode
-         * @return Returns 1 if characters are available at the serial port,
-         *         0 if no characters are available, and -1 if unsuccessful.
-         */
-        virtual std::streamsize showmanyc() override ;
-
+    
     private:
-
         /**
-         * @brief Forward declaration of the Implementation class folowing
-         *        the PImpl idiom.
+         * @brief The SerialStreamBuffer object that will be used by the
+         *        stream to communicate with the serial port.
          */
-        class Implementation;
+        std::unique_ptr<SerialStreamBuf> mIOBuffer {nullptr} ;
 
-        /**
-         * @brief Pointer to Implementation class instance.
-         */
-        std::unique_ptr<Implementation> mImpl;
-
-    } ; // class SerialStreamBuf
+    } ; // class SerialStream
 
 } // namespace LibSerial
