+++ test/SerialStreamUnitTests.h
@@ -33,10 +33,10 @@
 
 #pragma once
 
-#include "SerialPort.h"
-#include "SerialPortConstants.h"
-#include "SerialStream.h"
 #include "UnitTests.h"
+#include "libserial/SerialPort.h"
+#include "libserial/SerialPortConstants.h"
+#include "libserial/SerialStream.h"
 
 #include <gtest/gtest.h>
 #include <mutex>
